refactor(bridge): extract BridgeEventMapper from OrchestratorSession
